+++ runtime/src/jvmMain/kotlin/dev/zacsweers/metro/internal/Lock.jvm.kt
@@ -1,7 +1,0 @@
-// Copyright (C) 2026 Zac Sweers
-// SPDX-License-Identifier: Apache-2.0
-package dev.zacsweers.metro.internal
-
-import java.util.concurrent.locks.ReentrantLock
-
-internal actual typealias Lock = ReentrantLock